Add ability for patrons to opt into wage garnishment (#488)
A new page has been added which allows a patron to accept the
University Libraries' Lending Policy, which, if the user is eligible,
opts them into wage garnishment.

Currently, only faculty and staff are eligible to opt into this
agreement. Opting in sets the patron standing to 'OK' and sets the
GARNISH-DT (garnish date) to the current date.